[#1132708720621401] Fix counts during updates for column duplicates
Replace queries will only return the number of rows but not the number
of replacements within a row.
We therefore must use our custom counting method to get accurate counts
when fields have more than one occurrence of the URL.
Serialized will always run first during updates so only the left overs
are counted this way (which is what we want).